+++ modules/support/src/lib.rs
@@ -689,7 +689,7 @@ pub trait ForeignChainStateQuery<AccountId, Call> {
 	/// - who: Account that is requesting the query,
 	/// - length_bound: Size of encoded call, needed to bound call size
 	/// - dispatchable_call: Call to be dispatched on the condition of a foreign chain state
-	fn query_task(who: AccountId, length_bound: u32, dispatchable_call: Call) -> DispatchResult;
+	fn query_task(who: AccountId, length_bound: usize, dispatchable_call: Call) -> DispatchResult;
 
 	/// Cancels query, and refunds account the fee.
 	/// params:
